refactor and readme

+++ lib/modules/cloudservices/remoteconfig/cloudservices_remoteconfig.dart
@@ -1,8 +1,6 @@
-import 'dart:ffi';
-
 abstract class CloudServicesRemoteConfig {
   Future setDefaults(Map<String, dynamic> params);
-  Future fetch(Long cacheExpiration);
+  Future fetch(int cacheExpiration);
   String getString(String key);
   bool getBoolean(String key);
 }
